# docs/openapi-mobile-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-App Mobile API",
    "version": "v1",
    "description": "Machine-readable mobile-facing HTTP contract for Q-App v1."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "required": ["requestId"],
        "properties": {
          "requestId": { "type": "string" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code", "meta"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": {},
          "meta": { "$ref": "#/components/schemas/Meta" }
        }
      },
      "BusinessGroupRef": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
      "BusinessLocationRef": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
...
      "Shop": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "city",
          "countryCode",
          "addressLine1",
          "latitude",
          "longitude",
          "serviceStationsCount",
          "queueEnabled",
          "queuePaused",
          "industryType",
          "approvalStatus",
          "distanceKm",
          "queueLength",
          "estimatedWaitMin",
          "bestJoinScore",
          "bestJoinReason",
          "isFavorite",
          "reviewSummary"
        ],
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "city": { "type": "string" },
          "region": { "type": ["string", "null"] },
          "countryCode": { "type": "string" },
          "addressLine1": { "type": "string" },
          "postalCode": { "type": ["string", "null"] },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "publicDescription": { "type": ["string", "null"] },
          "serviceStationsCount": { "type": "integer" },
          "queueEnabled": { "type": "boolean" },
          "queuePaused": { "type": "boolean" },
          "industryType": { "type": "string" },
          "approvalStatus": { "type": "string" },
          "distanceKm": { "type": ["number", "null"] },
          "queueLength": { "type": "integer" },
          "estimatedWaitMin": { "type": "integer" },
          "bestJoinScore": { "type": "number" },
          "bestJoinReason": { "type": "string" },
          "isFavorite": { "type": "boolean" },
          "favoriteSince": { "type": "string", "format": "date-time" },
          "reviewSummary": { "$ref": "#/components/schemas/ReviewSummary" },
          "reviews": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicReview" }
          }
        }
      },
      "ReviewSummary": {
        "type": "object",
        "required": ["averageRating", "ratingCount"],
        "properties": {
          "averageRating": { "type": ["number", "null"] },
          "ratingCount": { "type": "integer" }
        }
      },
      "PublicReview": {
        "type": "object",
        "required": ["id", "rating", "comment", "createdAt", "customerName"],
        "properties": {
          "id": { "type": "string" },
          "rating": { "type": ["integer", "null"], "minimum": 1, "maximum": 5 },
          "comment": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "customerName": { "type": "string" }
        }
      },
      "FavoriteShopResult": {
        "type": "object",
        "required": ["slug", "isFavorite"],
        "properties": {
          "slug": { "type": "string" },
          "isFavorite": { "type": "boolean" }
        }
      },
...
      "Invitation": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "note",
          "createdAt",
          "expiresAt",
          "acceptedAt",
          "revokedAt",
          "businessGroup",
          "businessLocation"
        ],
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "role": { "type": "string" },
          "note": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "acceptedAt": { "type": ["string", "null"], "format": "date-time" },
          "revokedAt": { "type": ["string", "null"], "format": "date-time" },
          "businessGroup": { "$ref": "#/components/schemas/BusinessGroupRef" },
          "businessLocation": { "$ref": "#/components/schemas/BusinessLocationRef" },
          "inviteToken": { "type": "string" },
          "inviteUrl": { "type": "string" }
        }
      },
      "InvitationAcceptanceResult": {
        "type": "object",
        "required": ["invitation", "acceptedUser"],
        "properties": {
          "invitation": { "$ref": "#/components/schemas/Invitation" },
          "acceptedUser": {
            "type": "object",
            "required": ["id", "email", "firstName", "lastName"],
            "properties": {
              "id": { "type": "string" },
              "email": { "type": ["string", "null"] },
              "firstName": { "type": "string" },
              "lastName": { "type": ["string", "null"] }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Invalid payload.",
              "code": "BAD_REQUEST",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Resource not found.",
              "code": "NOT_FOUND",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": "Conflict.",
              "code": "CONFLICT",
              "meta": { "requestId": "request-id" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/preferences/favorites": {
      "get": {
        "tags": ["Discovery"],
        "operationId": "listFavoriteShops",
        "summary": "List favourite shops for an anonymous device",
        "parameters": [
          {
            "name": "X-QApp-Device-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "latitude", "in": "query", "schema": { "type": "number" } },
          { "name": "longitude", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "Favourite shop list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Shop" }
                    },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Shop" }
                    },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": [],
                  "items": [],
                  "meta": { "count": 0, "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/preferences/favorites/{shopSlug}": {
      "put": {
        "tags": ["Discovery"],
        "operationId": "addFavoriteShop",
        "summary": "Mark a shop as favourite for an anonymous device",
        "parameters": [
          {
            "name": "X-QApp-Device-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "shopSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Favourite saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/FavoriteShopResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": { "slug": "demo-barber", "isFavorite": true },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Discovery"],
        "operationId": "removeFavoriteShop",
        "summary": "Remove a favourite shop for an anonymous device",
        "parameters": [
          {
            "name": "X-QApp-Device-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "shopSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Favourite removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/FavoriteShopResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": { "slug": "demo-barber", "isFavorite": false },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/auth/invitations/{token}": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "getInvitationByToken",
        "summary": "Load invitation context by token",
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Invitation context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Invitation" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "id": "invite-id",
                    "email": "staff@shop.example",
                    "firstName": "Sam",
                    "lastName": "Operator",
                    "role": "STAFF_OPERATOR",
                    "note": "Front queue operations",
                    "createdAt": "2026-06-12T09:40:00.000Z",
                    "expiresAt": "2026-06-19T09:40:00.000Z",
                    "acceptedAt": null,
                    "revokedAt": null,
                    "businessGroup": {
                      "id": "group-id",
                      "name": "Fade Yard",
                      "slug": "q-app-demo-group"
                    },
                    "businessLocation": {
                      "id": "shop-id",
                      "name": "Fade Yard Soho",
                      "slug": "demo-barber"
                    }
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/invitations/{token}/accept": {
      "post": {
        "tags": ["Invitations"],
        "operationId": "acceptInvitation",
        "summary": "Accept invitation and create account access",
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "password": { "type": "string" }
                }
              },
              "example": {
                "firstName": "Sam",
                "lastName": "Operator",
                "password": "QappInviteStrong123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/InvitationAcceptanceResult" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "invitation": {
                      "id": "invite-id",
                      "email": "staff@shop.example",
                      "firstName": "Sam",
                      "lastName": "Operator",
                      "role": "STAFF_OPERATOR",
                      "note": "Front queue operations",
                      "createdAt": "2026-06-12T09:40:00.000Z",
                      "expiresAt": "2026-06-19T09:40:00.000Z",
                      "acceptedAt": "2026-06-12T09:45:00.000Z",
                      "revokedAt": null,
                      "businessGroup": {
                        "id": "group-id",
                        "name": "Fade Yard",
                        "slug": "q-app-demo-group"
                      },
                      "businessLocation": {
                        "id": "shop-id",
                        "name": "Fade Yard Soho",
                        "slug": "demo-barber"
                      }
                    },
                    "acceptedUser": {
                      "id": "user-id",
                      "email": "staff@shop.example",
                      "firstName": "Sam",
                      "lastName": "Operator"
                    }
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/ops/shops/{slug}/invitations": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "listInvitations",
        "summary": "List business invitations for one shop",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Invitation list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Invitation" }
                    },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Invitation" }
                    },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": [],
                  "items": [],
                  "meta": { "count": 0, "requestId": "request-id" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Invitations"],
        "operationId": "createInvitation",
        "summary": "Create a business invitation",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "role"],
                "properties": {
                  "email": { "type": "string" },
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "role": { "type": "string" },
                  "note": { "type": "string" }
                }
              },
              "example": {
                "email": "staff@shop.example",
                "firstName": "Sam",
                "lastName": "Operator",
                "role": "STAFF_OPERATOR",
                "note": "Front queue operations"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Invitation" },
                    "item": {},
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                },
                "example": {
                  "data": {
                    "id": "invite-id",
                    "email": "staff@shop.example",
                    "firstName": "Sam",
                    "lastName": "Operator",
                    "role": "STAFF_OPERATOR",
                    "note": "Front queue operations",
                    "createdAt": "2026-06-12T09:40:00.000Z",
                    "expiresAt": "2026-06-19T09:40:00.000Z",
                    "acceptedAt": null,
                    "revokedAt": null,
                    "businessGroup": {
                      "id": "group-id",
                      "name": "Fade Yard",
                      "slug": "q-app-demo-group"
                    },
                    "businessLocation": {
                      "id": "shop-id",
                      "name": "Fade Yard Soho",
                      "slug": "demo-barber"
                    },
                    "inviteToken": "<dev-token>",
                    "inviteUrl": "http://127.0.0.1:3000/invite/<dev-token>"
                  },
                  "item": {},
                  "meta": { "requestId": "request-id" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  }
}